+++ System/Hardware/Arduino/Protocol.hs
@@ -41,6 +41,7 @@ package (AnalogReport  p b)      = nonSysEx (REPORT_ANALOG_PIN p)   [if b then 1
 package (DigitalReport p b)      = nonSysEx (REPORT_DIGITAL_PORT p) [if b then 1 else 0]
 package (SetPinMode p m)         = nonSysEx SET_PIN_MODE            [fromIntegral (pinNo p), fromIntegral (fromEnum m)]
 package (DigitalPortWrite p l m) = nonSysEx (DIGITAL_MESSAGE p)     [l, m]
+package (AnalogPinWrite p l m)   = nonSysEx (ANALOG_MESSAGE p)      [l, m]
 package (SamplingInterval l m)   = sysEx    SAMPLING_INTERVAL       [l, m]
 package (Pulse p b dur to)       = sysEx    PULSE                   ([fromIntegral (pinNo p), if b then 1 else 0] ++ concatMap toArduinoBytes (word2Bytes dur ++ word2Bytes to))